+++ R/qenv-get_warnings.R
@@ -1,0 +1,54 @@
+#' Get the warnings of `qenv` object
+#'
+#' @param object (`qenv`)
+#'
+#' @return `character` containing warning information or `NULL` if no warnings
+#' @export
+#'
+#' @examples
+#' data_q <- new_qenv()
+#' data_q <- eval_code(new_qenv(), "iris_data <- iris")
+#' warning_qenv <- eval_code(
+#'   data_q,
+#'   bquote(p <- hist(iris_data[, .("Sepal.Length")], ff = ""))
+#' )
+#' cat(get_warnings(warning_qenv))
+#' @export
+setGeneric("get_warnings", function(object) {
+  # this line forces evaluation of object before passing to the generic
+  # needed for error handling to work properly
+  object
+
+  standardGeneric("get_warnings")
+})
+
+#' @rdname get_warnings
+#' @export
+setMethod("get_warnings", signature = c("qenv.error"), function(object) {
+  NULL
+})
+
+#' @rdname get_warnings
+#' @export
+setMethod("get_warnings", signature = c("qenv"), function(object) {
+  if (all(object@warnings == "")) {
+    return(NULL)
+  }
+
+  warning_output <- "Warnings:"
+  for (warn_idx in seq_along(object@warnings)) {
+    warn <- object@warnings[warn_idx]
+    if (warn != "") {
+      warning_output <- paste(
+        warning_output, "\n>", warn, "\nWhen running code:\n", paste(object@code[warn_idx], collapse = "\n")
+      )
+    }
+  }
+  paste0(warning_output, "\n\nTrace:\n", paste(get_code(object), collapse = "\n"))
+})
+
+#' @rdname get_warnings
+#' @export
+setMethod("get_warnings", "NULL", function(object) {
+  NULL
+})