+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.code 0.2.0.9000
 
+### Major breaking change
+* `chunks` have now been deprecated and will be removed from the package in a future release. The new `qenv` object should be used instead. See the new `qenv` vignette in the package for further details.
+
 # teal.code 0.2.0
 
 ### Miscellaneous